source/web_service/pyproject.toml
[tool.poetry]
name = "web-service"
version = "0.1.0"
description = ""
authors = ["Hector Pulido <hectorandrespp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.6"
djangorestframework = "^3.16.1"
drf-spectacular = "^0.28.0"
daphne = "^4.2.1"
channels = "^4.3.1"
psycopg2-binary = "^2.9.10"
requests = "^2.32.5"
dotenv = "^0.9.9"
websockets = "^15.0.1"
openai = "^1.107.0"

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**/*",
]


[tool.poetry.group.dev.dependencies]
pyrefly = "^0.32.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
